requirements.txt
django==2.2
django-rest-framework==3.9.2
requests==2.21.0